Ajouter le CSRF et Améliorer la gestion de quelques Erreurs des r2sultats de requêtes

+++ views/auteur/dashboard.php
@@ -1,5 +1,11 @@
 <?php
 session_start();
+
+
+if (empty($_SESSION['csrf_token'])) {
+    $_SESSION['csrf_token'] = bin2hex(random_bytes(32));
+}
+
 if ($_SESSION['role'] !== 'Auteur') {
     if ($_SESSION['role'] === 'Admin') {
         header("Location: ../admin/dashboard.php");
@@ -203,8 +209,8 @@ if ($_SESSION['role'] !== 'Auteur') {
 
                                 $cat = new Categorie();
                                 $result = $cat->distributeCategories((int) $_SESSION['id_user']);
-
-                                if (count($result) > 0) {
+                                // echo $result;
+                                if ($result) {
                                     foreach ($result as $row) {
                                         echo '
                                                 <div class="p-4 bg-gray-50 rounded-lg">
@@ -221,6 +227,8 @@ if ($_SESSION['role'] !== 'Auteur') {
                                                 </div>
                                             ';
                                     }
+                                }else{
+                                    echo '<h1 class="text-red-600 font-semibold text-lg">Vous n\'avez pas posté AUCUN ARTICLE !</h1>';
                                 }
 
                                 ?>
@@ -243,7 +251,7 @@ if ($_SESSION['role'] !== 'Auteur') {
                             $auteur = new Auteur();
                             $articles = $auteur->recentArticles((int)$_SESSION['id_user']);
 
-                            if (count($articles) > 0) {
+                            if ($articles) {
                                 foreach ($articles as $article) {
                                     echo '
                                             <div
@@ -267,6 +275,8 @@ if ($_SESSION['role'] !== 'Auteur') {
                                             </div>
                                         ';
                                 }
+                            }else{
+                                echo '<h1 class="text-red-600 font-semibold text-lg">Vous n\'avez pas posté AUCUN ARTICLE !</h1>';
                             }
                             ?>
                         </div>
@@ -284,36 +294,39 @@ if ($_SESSION['role'] !== 'Auteur') {
                 $articles = $auteur->ownArticles((int) $_SESSION['id_user']);
 
                 // print_r($articles);
-                
-                foreach ($articles as $art) {
-                    echo '<article class="relative bg-white shadow-md rounded-md">';
-                    echo '<div>
-                                <img src="../../assets/img/default-image.png" class="rounded-t-md" alt="Couverture de l\'Article">
-                            </div>';
-                    echo '<div class="p-4">';
-                    echo '<p class="text-gray-800 font-medium text-sm">' . $art['date_publication'] . ' •</p>';
-                    echo '<div class="pt-5">
-                                    <a href="#"><h1 class="text-gray-900 font-semibold text-xl mb-3">' . $art['titre'] . '</h1></a>
-                                    <p class="text-gray-700 font-medium text-md">' . substr($art['contenu'], 0, 100) . '...</p>
+                if($articles){
+                    foreach ($articles as $art) {
+                        echo '<article class="relative bg-white shadow-md rounded-md">';
+                        echo '<div>
+                                    <img src="../../assets/img/default-image.png" class="rounded-t-md" alt="Couverture de l\'Article">
+                                </div>';
+                        echo '<div class="p-4">';
+                        echo '<p class="text-gray-800 font-medium text-sm">' . $art['date_publication'] . ' •</p>';
+                        echo '<div class="pt-5">
+                                        <a href="#"><h1 class="text-gray-900 font-semibold text-xl mb-3">' . $art['titre'] . '</h1></a>
+                                        <p class="text-gray-700 font-medium text-md">' . substr($art['contenu'], 0, 100) . '...</p>
+                                    </div>
+                                    <div class="flex justify-end items-center gap-5 mt-5">
+                                        <a href="#">
+                                            <button type="button" class="py-2 px-5 rounded-sm text-white bg-blue-500 text-sm duration-500 hover:bg-blue-700">Modifier</button>
+                                        </a>
+                                        <a href="../../actions/deleteArticle.php?id=' . $art['id_article'] . '">
+                                            <button type="button" class="py-2 px-5 rounded-sm text-white bg-red-500 text-sm  duration-500 hover:bg-red-700">Supprimer</button>
+                                        </a>
+                                    </div>
                                 </div>
-                                <div class="flex justify-end items-center gap-5 mt-5">
-                                    <a href="#">
-                                        <button type="button" class="py-2 px-5 rounded-sm text-white bg-blue-500 text-sm duration-500 hover:bg-blue-700">Modifier</button>
-                                    </a>
-                                    <a href="../../actions/deleteArticle.php?id=' . $art['id_article'] . '">
-                                        <button type="button" class="py-2 px-5 rounded-sm text-white bg-red-500 text-sm  duration-500 hover:bg-red-700">Supprimer</button>
-                                    </a>
-                                </div>
-                            </div>
-                            <p class="absolute top-2 right-2 bg-white bg-opacity-85 py-1 px-3 rounded-md text-xs">' . $art['nom_categorie'] . '</p>';
-                            if ($art['etat'] == 'Accepté') {
-                                echo '<span class="absolute top-2 left-2 text-xs bg-green-100 text-green-600 px-2 py-1 rounded-full">' . $art['etat'] . '</span>';
-                            } else if ($art['etat'] == 'En Attente') {
-                                echo '<span class="absolute top-2 left-2 text-xs bg-yellow-100 text-yellow-600 px-2 py-1 rounded-full">' . $art['etat'] . '</span>';
-                            } else {
-                                echo '<span class="absolute top-2 left-2 text-xs bg-red-100 text-red-600 px-2 py-1 rounded-full">' . $art['etat'] . '</span>';
-                            }
-                        echo '</article>';
+                                <p class="absolute top-2 right-2 bg-white bg-opacity-85 py-1 px-3 rounded-md text-xs">' . $art['nom_categorie'] . '</p>';
+                                if ($art['etat'] == 'Accepté') {
+                                    echo '<span class="absolute top-2 left-2 text-xs bg-green-100 text-green-600 px-2 py-1 rounded-full">' . $art['etat'] . '</span>';
+                                } else if ($art['etat'] == 'En Attente') {
+                                    echo '<span class="absolute top-2 left-2 text-xs bg-yellow-100 text-yellow-600 px-2 py-1 rounded-full">' . $art['etat'] . '</span>';
+                                } else {
+                                    echo '<span class="absolute top-2 left-2 text-xs bg-red-100 text-red-600 px-2 py-1 rounded-full">' . $art['etat'] . '</span>';
+                                }
+                            echo '</article>';
+                    }
+                }else{
+                    echo '<h1 class="col-span-2 text-red-600 font-semibold text-2xl">Vous n\'avez pas posté AUCUN ARTICLE !</h1>';
                 }
 
                 ?>
@@ -329,6 +342,7 @@ if ($_SESSION['role'] !== 'Auteur') {
                         </h2>
                     </div>
                     <form method="POST" action="../../actions/addArticle.php" id="addArticleForm" class="mt-8 space-y-6">
+                        <input type="hidden" name="csrf_token" value="<?php echo $_SESSION['csrf_token']; ?>">
                         <div class="rounded-md shadow-sm flex flex-col gap-5">
                             <div>
                                 <label for="titre" class="sr-only">Titre</label>
